requirements.txt
aiogram==3.13.1
aiomysql==0.2.0
python-dotenv==1.0.1